api/requirements.txt
flask
flask-cors
requests
pypco
flask_sqlalchemy
pyjwt
